% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{annotateByKnownPAC}
\alias{annotateByKnownPAC}
\title{Annotate a PACdataset with known PACdatasets}
\usage{
annotateByKnownPAC(pacds, knownPACdss, labels, d=50, verbose=TRUE)
}
\arguments{
\item{pacds}{a query PACdataset.}

\item{knownPACdss}{a PACdataset or a list of PACdatasets.}

\item{labels}{a character vector storing column names added to pacds.
Each PACdataset in knownPACdss will add two columns ({labels[i]}_ovp, _dist). The length of labels must be 1 or the same as knowPACdss (if it is a list).}

\item{verbose}{print information.}
}
\value{
a PACdataset with columns ({labels[i]}_ovp, _dist) added.
The _ovp column is 0/1 denoting whether a PAC in pacds is overlapping with knowPACdss.
The _dist column is NA (no overlapping) or a integer denoting the min distance between a PAC in pacds and the overlapped PAC in knownPACdss.
}
\description{
annotateByKnownPAC returns overlapping status and min distance with given PACdatasets.
}
\details{
This function is used to validate a PACdataset with other known PAC datasets and calculates the min distance between two PACs.
}
\examples{
data(PACds); pacds=PACds
knownPACdss=list(pacds[1:500], pacds[500:1000])
p2=annotateByKnownPAC(pacds, knownPACdss, labels=c('test1','test2'))
summary(p2@anno$test2_dist)
}
\seealso{
[findOvpPACds()] to find overlapping between two PACdatasets.

Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{countOvpPACds}()},
\code{\link{ext3UTRPACds}()},
\code{\link{findOvpPACds}()},
\code{\link{get3UTRAPAds}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{samplePACds}()},
\code{\link{scPACds}},
\code{\link{subsetPACds}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
