% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{asAnyMatrix}
\alias{asAnyMatrix}
\title{Convert to matrix or dgCMatrix}
\usage{
asAnyMatrix(dat, forceSparse = FALSE)
}
\arguments{
\item{dat}{a data.frame/matrix/dgCMatrix or any object that can be converted to matrix using as.matrix.}

\item{forceSparse}{if TRUE, then force the dat to sparse matrix (dgCMatrix)}
}
\value{
A matrix or dgCMatrix which can be used as the count slot of a PACdataset.
}
\description{
asAnyMatrix convert data.frame/matrix/dgCMatrix to matrix or dgCMatrix.
For PACdataset@counts in previous movAPA, the data type is data.frame. Please call asAnyMatrix(PACds@counts) to make it compatible to new movAPA.
If dat is data-sparse or type-sparse, then return dgCMatrix
if dat is data.frame, then return matrix or dgCMatrix (when forceSparse or data-sparse)
}
