% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{countOvpPACds}
\alias{countOvpPACds}
\title{Count overlapping between PACdatasets}
\usage{
countOvpPACds(pacds1, pacds2, d=50, pats1=1, pats2=1)
}
\arguments{
\item{pacds1}{a PACdataset.}

\item{pacds2}{another PACdataset.}

\item{mode1}{use coord or UPA_start/UPA_end for pacds1, default is point (using coord). If mode1=region, then use UPA_start/UPA_end.}

\item{mode2}{same as mode1 except for pacds2.}

\item{d}{distance to count overlapping, default is 0 (no gap between qry and sbj).}

\item{pats1}{cutoffs to filter pacds1 by total tag number of all samples. If pats1<=1, then do not filter.
Multiple cutoffs can be set, like pats1=c(10, 20, 30).}

\item{pats2}{same as pats1 but to filter pacds2.}
}
\value{
a dataframe with 8 columns: PAT1, PAT2, Total1, Total2, Ovp1, Ovp2, Ovp1Pct, Ovp2Pct.
Row number is the combination number of pats1 and pats2.
}
\description{
countOvpPACds counts overlapping between PACdatasets.
}
\details{
This function is used to compare PA coordinates of two PACdataset objects.
}
\examples{
data(PACds)
pacds1=PACds
pacds2=PACds
pacds2@anno$coord=pacds2@anno$coord+sample(-10:10, 1)
## Count exactly the same positions between two pacds without filtering
countOvpPACds(pacds1, pacds2, d=0)
## count overlapping with different cutoffs
countOvpPACds(pacds1, pacds2, d=50, pats1=1, pats2=seq(0,20,10))
## count overlapping with different cutoffs, comparing ds1's coord ~ ds2's region
countOvpPACds(pacds1, pacds2, mode1='point', mode2='region', d=50, pats1=1, pats2=seq(0,20,10))
}
\seealso{
[findOvpPACds()] to get an overlapped PACdataset.

Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByKnownPAC}()},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{ext3UTRPACds}()},
\code{\link{findOvpPACds}()},
\code{\link{get3UTRAPAds}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{samplePACds}()},
\code{\link{scPACds}},
\code{\link{subsetPACds}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
