% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{createPACdataset}
\alias{createPACdataset}
\title{Create a PACdataset from counts data}
\usage{
createPACdataset(counts, anno, colData, supp, forceSparse = F, verbose = TRUE)
}
\arguments{
\item{counts}{PAC expression or ratio data, can be data.frame/matrix/dgCMatrix.
If the matrix is in character mode, then will be converted into numeric mode.}

\item{anno}{detailed annotation information for PAs, with the same row number as the counts slot.}

\item{colData}{sample defination. Each row is one experiment, and the column is the sample group.}

\item{supp}{a list storing additional data, like stopCodon.}

\item{forceSparse}{whether to force the counts matrix as sparse matrix. By default the function determines whether the matrix is sparse (if it contains more than 50 percentage of 0s and ncol is greater than 10). But set forceSparse=TRUE can skip the sparse check.}

\item{verbose}{TRUE to print message.}
}
\value{
A \code{\link{PACdataset}} object
}
\description{
createPACdataset create a PACdataset from counts and anno matrix/data.frame.
This function is particularly suitable for reading PACdatasets from single-cell data with sparse counts table and a separate meta data.
}
\examples{
data(PACds)
p1=createPACdataset(counts=PACds@counts, anno=PACds@anno)
anno=PACds@anno; rownames(anno)=NULL
p2=createPACdataset(counts=PACds@counts, anno=anno)
p3=createPACdataset(counts=PACds@counts, anno=anno, forceSparse=TRUE)
}
\seealso{
\code{\link{readPACds}} to read PACdataset from file or data.frame.

Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{get3UTRAPAds}()},
\code{\link{get3UTRAPApd}()},
\code{\link{length}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{scPACds}},
\code{\link{subscript_operator}},
\code{\link{summary}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
