% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\docType{data}
\name{demo_peaks}
\alias{demo_peaks}
\title{A demo peaks list of human}
\format{
A list including peaks.meta and peaks.demo.
}
\source{
One result from scAPAtrap.
}
\usage{
demo_peaks
}
\description{
demo_peaks stores a list containing 5w peaks in 2538 cells, including peaks.meta and peaks.demo. This file was obtained by scAPAtrap.
}
\keyword{datasets}
