% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{findOvpPACds}
\alias{findOvpPACds}
\title{Find overlapping PACdataset}
\usage{
countOvpPACds(qryPACds, sbjPACds, d=50, qryPAT=1, sbjPAT=1, verbose=TRUE)
}
\arguments{
\item{qryPACds}{a query PACdataset, will subset and return the qryPACds.}

\item{sbjPACds}{a subject PACdataset.}

\item{qryMode}{use coord or UPA_start/UPA_end for qryPACds. Default is point (using coord), if qryMode=region, then use UPA_start/UPA_end.}

\item{sbjMode}{same as qryMode except for sbjPACds.}

\item{d}{distance to count overlapping, default is 0 (no gap between qry and sbj).}

\item{qryPAT}{cutoff to filter qryPACds. If sbjPAT<=1, then do not filter.}

\item{sbjPAT}{same as qryPAT but to filter sbjPACds.}

\item{returnOvp}{default is TRUE; if TRUE then return overlapping qryPACds.}

\item{returnNonOvp}{default is FALSE; if TRUE then return nonOverlapping qryPACds.
If both returnOvp=T and returnNonOvp=TRUE, then return a list(ovp, nonOvp)}

\item{verbose}{print information.}
}
\value{
a subset PACdataset of qryPACds.
}
\description{
findOvpPACds returns overlapping PACdataset.
}
\details{
This function is used to subset one PACds with another PACds.
}
\examples{
data(PACds)
pacds1=PACds
pacds2=PACds
pacds2@anno$coord=pacds2@anno$coord+sample(-10:10, 1)
## return overlapped qryPACds without filtering, comparing coord~coord
pd=findOvpPACds(qryPACds=pacds1, sbjPACds=pacds2, d=0, qryPAT=1, sbjPAT=1)
## return overlapped qryPACds with filtering, comparing coord~coord
pd=findOvpPACds(qryPACds=pacds1, sbjPACds=pacds2, d=50, qryPAT=1, sbjPAT=10)
## return overlapped qryPACds comparing qry's coord ~ sbj's UPA_start/end
pd=findOvpPACds(qryPACds=pacds1, sbjPACds=pacds2, d=0, qryMode='point', sbjMode='region')
}
\seealso{
[countOvpPACds()] to count overlapping between two PACdatasets.

Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByKnownPAC}()},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{countOvpPACds}()},
\code{\link{ext3UTRPACds}()},
\code{\link{get3UTRAPAds}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{samplePACds}()},
\code{\link{scPACds}},
\code{\link{subsetPACds}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
