% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{get3UTRAPAds}
\alias{get3UTRAPAds}
\title{Get 3'UTR APA PACdataset}
\usage{
get3UTRAPAds(
  pacds,
  sortPA = TRUE,
  choose2PA = NULL,
  avgPACtag = 0,
  avgGeneTag = 0,
  clearPAT = 0
)
}
\arguments{
\item{pacds}{A PACdataset object.}

\item{sortPA}{If TRUE, then order PACs by the respective 3'UTR length in each gene.}

\item{choose2PA}{specify whether and how to choose two PACs when there are >2 PACs.
The value can be NULL (use all PACs), PD (choose only proximal and distal sites),
farest (choose two PACs that are with the longest distance), most (choose two PACs with the most abundance).}

\item{avgPACtag}{if >0, then to filter by PAC tag num, >=avgPACtag, see subsetPACds().}

\item{avgGeneTag}{if >0, then to filter by PAC tag num, >=avgGeneTag, see subsetPACds().}

\item{clearPAT}{if >0, then to filter by clearPAT, see subsetPACds().}
}
\value{
A PACdataset with only 3'UTR APA sites. If there is no result, return an empty PACdataset with 0-row @anno and @counts.
}
\description{
get3UTRAPAds subset a PACdataset to get all PACs of genes with 3'UTR APA sites.
}
\details{
Genes with 3'UTR APA sites are genes with multiple PACs in its 3'UTR.
If the column toStop is not in PACds@anno, then will calculate the 3'UTR length first.
First filter by paramters like avgPACtag, etc., and then by choose2PA.
}
\examples{
data(PACds)
## Get all 3'UTR APA
pp1=get3UTRAPAds(PACds,sortPA=TRUE); summary(pp1); is3UTRAPAds(pp1); is3UTRAPAds(PACds)
## Get proximal distal PA
pp1=get3UTRAPAds(PACds, sortPA=TRUE, choose2PA='most');
summary(pp1); is3UTRAPAds(pp1); is3UTRAPAds(PACds); head(pp1@anno)
pp1@anno[pp1@anno$gene=='PH02Gene00018',]
}
\seealso{
Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{createPACdataset}()},
\code{\link{get3UTRAPApd}()},
\code{\link{length}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{scPACds}},
\code{\link{subscript_operator}},
\code{\link{summary}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
