% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{get3UTRAPApd}
\alias{get3UTRAPApd}
\title{Get proximal and distal 3'UTR APAs}
\usage{
get3UTRAPApd(
  pacds,
  minDist = 50,
  maxDist = 10000,
  minRatio = 0.05,
  fixDistal = FALSE,
  addCols = "pd"
)
}
\arguments{
\item{pacds}{a PACdataset, if it contains non-3UTR PAs, will run get3UTRAPAds first to subset 3UTR PAs.}

\item{minDist}{the distance of the proximal and distal PA should be in [minDist, maxDist].}

\item{maxDist}{the distance of the proximal and distal PA should be in [minDist, maxDist].
It is recommended to use \code{\link{getNearbyPAdist}} to estimate an appropriate distance.}

\item{minRatio}{If minRatio>0, then will remove PAs with ratio<minRatio before getting P/D PAs.}

\item{fixDistal}{TRUE means that the most distal PA is fixed and then only searches for the proximal PA for that distal PA.}

\item{addCols}{If not NULL, then four columns would be added to PACds@anno, inlucidng <addCols>Which/Score/Ratio/Dist. Among these columns, the 'which' column denotes P or D of each PA.}
}
\value{
A PACdataset with only 3'UTR P-D APA sites. If there is no result, return an empty PACdataset with 0-row @anno and @counts.
}
\description{
get3UTRAPApd provides a smarter way to get proximal and distal 3'UTR APAs then get3UTRAPAds.
Criteria for a proximal and distal pair in a gene: first select APA pairs from all APA pairs that meet the dist and ratio criteria; then the final P/D pair is the one with the highest count sum and the closest distance.
Note that there is no filtering of PA expression levels here, which can be performed before get3UTRAPApd using subsetPACds().
}
\examples{
data(PACds)
## to get a glance of the distance between nearby 3UTR pAs
getNearbyPAdist(PACds, ftrs='3UTR', within='ftr', return='null')
pd1=get3UTRAPApd(PACds, minDist=50, maxDist=1000, minRatio=0.05, fixDistal=F)
pd2=get3UTRAPApd(PACds, minDist=50, maxDist=1000, minRatio=0.05, fixDistal=T)
g=PACds@anno$gene[!(pacds@anno$gene \%in\% pd1@anno$gene)]
subsetPACds(pd1, gene=g)
subsetPACds(pd2, gene=g)
}
\seealso{
\code{\link{subsetPACds}} \code{\link{get3UTRAPAds}} \code{\link{movAPAindex}}

Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{createPACdataset}()},
\code{\link{get3UTRAPAds}()},
\code{\link{length}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{scPACds}},
\code{\link{subscript_operator}},
\code{\link{summary}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
