% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{getVarGrams}
\alias{getVarGrams}
\title{Get actual grams for an abbreviation}
\usage{
getVarGrams(grams)
}
\arguments{
\item{grams}{a character string (e.g., V1, v1, MOUSE, mouse, mm10, mm) or a string vector.
V1 means AATAAA and its variants; MOUSE means hexamers of polyA signals in mouse.
Mouse signals are obtained from https://polyasite.unibas.ch/atlas#2, which reside in a region of 60 nt upstream to 10 nt downstream of one of the poly(A) sites of a cluster.}
}
\value{
A vector of grams.
}
\description{
getVarGrams returns a vector of grams given an abbreviation.
}
\details{
This function is used to get common grams.
}
\examples{
## get AATAAA and its variants
getVarGrams('v1')
## get mouse hexamers
getVarGrams('mm')
## not doing anything
getVarGrams(c('AATAAA','AATT'))
}
\seealso{
Other APA signal functions: 
\code{\link{annotateByPAS}()},
\code{\link{faFromPACds}()},
\code{\link{kcount}()},
\code{\link{plotATCGforFAfile}()},
\code{\link{plotSeqLogo}()}
}
\concept{APA signal functions}
