% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\docType{class}
\name{heatmapResults-class}
\alias{heatmapResults-class}
\alias{heatmapResults}
\title{heatmapResult object}
\description{
heatmapResult is an S4 class to represent a list with padj, value and colData. Normally it is from getHeatmapRes().
}
\details{
The slot of padj or value is a data frame with each row being gene/PA, each column being a condpair or cond (if DEXSeq, then it is the exonUsage matrix).
}
\section{Slots}{

\describe{
\item{\code{padj}}{a data frame recording the pvalue of each gene or PA in each condition.}

\item{\code{value}}{a data frame with the same dimension as padj.}

\item{\code{colData}}{stores padj/value matrix's annotation info. Its row names are the column names of the padj matrix.
Normally, there are two columns (cond1 and cond2) in colData.}
}}

\seealso{
Other movRes objects: 
\code{\link{movAPASwitchRes-class}},
\code{\link{movDEGeneRes-class}},
\code{\link{movIndexDiffRes-class}},
\code{\link{movRes-class}},
\code{\link{movUTRTrendRes-class}}
}
\concept{movRes objects}
