% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{length}
\alias{length}
\title{Length of a PACdataset}
\usage{
\S4method{length}{PACdataset}(x)
}
\arguments{
\item{x}{a PACdataset}
}
\value{
The length, i.e., the number of pAs, i.e., the number of rows in slot counts or anno.
}
\description{
Length of a PACdataset
}
\seealso{
Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{createPACdataset}()},
\code{\link{get3UTRAPAds}()},
\code{\link{get3UTRAPApd}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{scPACds}},
\code{\link{subscript_operator}},
\code{\link{summary}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
