% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{makeExamplePACds}
\alias{makeExamplePACds}
\title{Make an example PACdataset}
\usage{
makeExamplePACds(n = 1000, seed = 123, forceSparse = F)
}
\arguments{
\item{n}{number of PACs.}

\item{seed}{random seed.}

\item{forceSparse}{whether to make the counts table as a sparse matrix, otherwise is a normal matrix.}
}
\value{
A PACdataset of Arabidopsis with Chr1..Chr5. There are six columns in the anno slot.
}
\description{
makeExamplePACds returns an example PACdataset of Arabidopsis.
}
\examples{
makeExamplePACds(n=1000)
makeExamplePACds(n=1000, seed=456, forceSparse=T)
}
\seealso{
Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{createPACdataset}()},
\code{\link{get3UTRAPAds}()},
\code{\link{get3UTRAPApd}()},
\code{\link{length}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{scPACds}},
\code{\link{subscript_operator}},
\code{\link{summary}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
