% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\docType{package}
\name{movAPA-package}
\alias{movAPA}
\alias{movAPA-package}
\title{movAPA: movAPA: Modeling and Visualization of dynamics of Alternative PolyAdenylation}
\description{
This package provides functions for modeling and visualization of dynamics of alternative polyadenylation across different experimental conditions.
}
\details{
The movAPA package is a comprehensive package for MOdeling and Visualization of dynamics of Alternative PolyAdenylation across biological samples in different species.
Functions of movAPA include pre-processing of poly(A) site clusters (PACs), poly(A) signal analysis, differential analyses, APA site switching analyses, etc.
The package is not species-specific, so it can be used for any species (providing qualified GFF annotation and genome references).
}
\references{
Zhu, S., et al. (2020) PlantAPAdb: A Comprehensive Database for Alternative Polyadenylation Sites in Plants, Plant Physiol., 182, 228-242.
Wu, X., et al. (2011) Genome-wide landscape of polyadenylation in Arabidopsis provides evidence for extensive alternative polyadenylation, Proc. Natl. Acad. Sci. USA, 108, 12533-12538.
}
