% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\docType{class}
\name{movAPASwitchRes-class}
\alias{movAPASwitchRes-class}
\title{movAPASwitchRes object}
\description{
movAPASwitchRes is an S4 class to represent APA site switching results among conditions.
}
\details{
The movAPASwitchRes object is obtained by calling movAPAswitch function.
movAPASwitchRes is for two PACs; movUTRTrendRes is for two or more PACs (dependent on the method used); movDEPACRes is only for one PAC.
}
\section{Slots}{

\describe{
\item{\code{fullList}}{records the full information of APA site switching results,
including 3' UTR length, poly(A) sites expression (PA1 and PA2), etc.}
}}

\seealso{
\code{\link{movAPAswitch}} for APA site switching analysis on a PACdataset.

Other movRes objects: 
\code{\link{heatmapResults-class}},
\code{\link{movDEGeneRes-class}},
\code{\link{movIndexDiffRes-class}},
\code{\link{movRes-class}},
\code{\link{movUTRTrendRes-class}}
}
\concept{movRes objects}
