% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{movDEGene}
\alias{movDEGene}
\title{Get DE genes}
\usage{
movDEGene(PACds, group, method='DESeq2', minSumPAT=10)
}
\arguments{
\item{PACds}{a PACdataset.}

\item{group}{a sample group name to get all conditions under the group.}

\item{method}{only DESeq2 is allowed for now}

\item{minSumPAT}{filter genes with total PAT number across all conditions >=minSumPAT}
}
\value{
An object of movDEGeneRes, with slots[group, conds, pairwise=list[value=log2FC_<cond1.cond2>, padj=padj_<cond1.cond2>]]
}
\description{
movDEGene first summerizes all PACs in a gene to get gene expression level, and then calls DESeq2 for DE analysis.
It performs DE analyses for all condition pairs of the given group. If you only need a specific condition pair, call
subsetPACds() first to filter used conditions.
}
\examples{
data(PACds)
## Detect DE genes for two conditions.
## Subset two conditions first.
pacds=subsetPACds(PACds, group='group',cond1='anther', cond2='embryo')
## Detect DE genes using DESeq2 method, only genes with total read counts in all samples >=50 are used.
DEgene=movDEGene(PACds=pacds, method='DESeq2', group='group', minSumPAT=50)
## Detect DE genes in all pairwise conditions.
DEgene=movDEGene(PACds=PACds, method='DESeq2', group='group', minSumPAT=50)
}
\seealso{
Other comparison functions: 
\code{\link{movAPAindexDiff}()},
\code{\link{movAPAindex}()},
\code{\link{movAPAswitch}()},
\code{\link{movDEPAC}()},
\code{\link{movPAindex}()},
\code{\link{movUTRtrend}()},
\code{\link{plotCummPAindex}()}
}
\concept{comparison functions}
