% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\docType{class}
\name{movDEGeneRes-class}
\alias{movDEGeneRes-class}
\title{movDEGeneRes object}
\description{
An S4 class to represent DE gene results among conditions.
}
\details{
The movDEGeneRes object is obtained by calling movDEGene function.
}
\seealso{
[movDEGene()] for DE gene analysis on a PACdataset.

Other movRes objects: 
\code{\link{heatmapResults-class}},
\code{\link{movAPASwitchRes-class}},
\code{\link{movDEPACRes-class}},
\code{\link{movIndexDiffRes-class}},
\code{\link{movRes-class}},
\code{\link{movUTRTrendRes-class}}
}
\concept{movRes objects}
