% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{movDEPAC}
\alias{movDEPAC}
\title{Get DE PACs}
\usage{
movDEPAC(PACds, group, method = "DESeq2", minSumPAT = 10, chisqPadjust = TRUE)
}
\arguments{
\item{PACds}{a PACdataset.}

\item{group}{a sample group name to get all conditions under the group.}

\item{method}{can be DEXseq, DESeq2, chisq, ignoring cases.
chisq is as in Shulman 2019 et al, which first performs chisq.test for all PACs in a gene to filter sig. genes and then performs chisq.test for each PAC.
If there are multiple reps, then will pool reps first before chisq.}

\item{minSumPAT}{filter PACs with total PAT number across all conditions >=minSumPAT}

\item{chisqPadjust}{if T then output adjust pvalues instead of raw pvalues.}
}
\value{
An object of movDEPACRes, including slots [group, conds,
                                 pairwise=list[value=log2FC_<cond1.cond2> for DESeq2/DEXSeq; 1-genePv for Chisq,
                                 padj=padj_<cond1.cond2>],
                                 PACusage(if method=DEXseq)]
}
\description{
movDEPAC calls DESeq2/DEXSeq or uses chisq.test for DE analysis.
It performs DE analyses for all condition pairs of the given group. If you only need a specific condition pair, call
subsetPACds first to filter used conditions.
}
\examples{
data(PACds)
## Detect DE PACs in all pairwise conditions using DESeq2 method,
## only PACs with total read counts in all samples >=20 are used.
DEPAC=movDEPAC(PACds, method='DESeq2', group='group', minSumPAT=20)

## Detect DE PACs using DEXseq method (may be slow).
DEXPAC=movDEPAC(PACds, method='DEXseq', group='group', minSumPAT=20)

## Detect DE PACs using chisq-PACdsPAS method.
DEqPAC=movDEPAC(PACds, method='chisq', group='group', minSumPAT=20)
}
\seealso{
Other comparison functions: 
\code{\link{movAPAindexDiff}()},
\code{\link{movAPAindex}()},
\code{\link{movAPAswitch}()},
\code{\link{movDEGene}()},
\code{\link{movPAindex}()},
\code{\link{movUTRtrend}()},
\code{\link{plotCummPAindex}()}
}
\concept{comparison functions}
