% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\docType{class}
\name{movIndexDiffRes-class}
\alias{movIndexDiffRes-class}
\title{movIndexDiffRes object}
\description{
movIndexDiffRes is an S4 class to represent results of APA index difference among conditions.
}
\details{
The movIndexDiffRes object is obtained by calling movAPAindexDiff function.
}
\section{Slots}{

\describe{
\item{\code{fullList}}{records the full information of APA index difference results.}
}}

\seealso{
[movAPAindexDiff()] for APA index difference analysis on a PACdataset.

Other movRes objects: 
\code{\link{heatmapResults-class}},
\code{\link{movAPASwitchRes-class}},
\code{\link{movDEGeneRes-class}},
\code{\link{movDEPACRes-class}},
\code{\link{movRes-class}},
\code{\link{movUTRTrendRes-class}}
}
\concept{movRes objects}
