% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{movPAindex}
\alias{movPAindex}
\title{Calculate PAC index for each PAC in each condition.}
\usage{
movPAindex(PACds, method, shan.ratio = FALSE)
}
\arguments{
\item{PACds}{a PACDataset object.}

\item{method}{should be "shannon","ratio", "geo".
\itemize{
\item{"shannon"}: tissue specificity;
\item{"ratio"}: ratio for APA sites (single PA genes are removed first);
\item{"geo"}: PUI by geometry mean (=log2(pA/all_PAs_geo_mean)), as in Shulman 2019 et al (single PA genes are removed first).
}}

\item{shan.ratio}{Pass to specificityByShannon.}
}
\value{
A matrix of PA index.
\itemize{
\item{shannon}: return a dataframe of [H (overall TS), Q_min (min TS), Q_min_tissue, Q for each tissue], in which smaller H or Q means higher tissue-specificity.
 If PA=0 in one sample, then the respective Qval is NA.
\item{"geo/ratio"}: return a dataframe of the same row# and col# of the PACds@counts.
}
}
\description{
movPAindex calculates index of each PAC, while movAPAindex calculates index for APA genes.
method=ratio/geo is only allowed for APA genes, so these indices are only calculated for PACs in APA genes.
}
\examples{
data(PACds)
PAindex=movPAindex(PACds, method='ratio')
PAindex=movPAindex(PACds, method='geo')
PAindex=movPAindex(PACds, method='shannon')
}
\seealso{
Other comparison functions: 
\code{\link{movAPAindexDiff}()},
\code{\link{movAPAindex}()},
\code{\link{movAPAswitch}()},
\code{\link{movDEGene}()},
\code{\link{movDEPAC}()},
\code{\link{movUTRtrend}()},
\code{\link{plotCummPAindex}()}
}
\concept{comparison functions}
