% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\docType{class}
\name{movRes-class}
\alias{movRes-class}
\title{movRes object}
\description{
movRes is an S4 class to represent comparison results among conditions by mov-like funcions (movDEPAC etc.).
}
\details{
All other movRes classes, such as movDEGeneRes, movDEPACRes, movUTRTrendRes,
movAPASwitchRes, movIndexDiffRes are generic from movRes. Use movSelect, movStat,
and movViz to process the movRes class.
}
\section{Slots}{

\describe{
\item{\code{type}}{type of the movRes object, such as DEGENE, DEPAC.}

\item{\code{method}}{define the method used for obtaining the movRes object, like DE, DEX, lineartrend, etc.}

\item{\code{group}}{group label of the samples.}

\item{\code{conds}}{a data frame with two columns specifying the two conditions (cond1 and cond2) in the sample group for the comparison analysis.
Rownames of the conds data frame are like "cond1.cond2".}

\item{\code{pairwise}}{a list with two elements, padj and value, which are normally obtained from getHeatmapRes().}
}}

\seealso{
\code{\link{movSelect}} for select movRes, \code{\link{movStat}} for making statistics, \code{\link{movViz}} to visualize a gene.

Other movRes objects: 
\code{\link{heatmapResults-class}},
\code{\link{movAPASwitchRes-class}},
\code{\link{movDEGeneRes-class}},
\code{\link{movIndexDiffRes-class}},
\code{\link{movUTRTrendRes-class}}
}
\concept{movRes objects}
