% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{movRes2heatmapResults}
\alias{movRes2heatmapResults}
\alias{movRes2heatmapResults,movRes-method}
\title{Convert movRes object to heatmapResults object}
\usage{
movRes2heatmapResults(aMovRes)

\S4method{movRes2heatmapResults}{movRes}(aMovRes)
}
\arguments{
\item{aMovRes}{a movRes object}
}
\value{
A heatmapResults object.
}
\description{
movRes2heatmapResults converts movRes object to heatmapResults object,
so that then can use subsetHeatmap functions. This function is mainly used by statGeneralHeat().
}
\section{Methods (by class)}{
\itemize{
\item \code{movRes2heatmapResults(movRes)}: for a movRes object

}}
