% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{movStat}
\alias{movStat}
\alias{movStat,movRes-method}
\alias{movStat,PACdataset-method}
\title{Make statistics for a object}
\usage{
movStat(object, ...)

\S4method{movStat}{movRes}(object, padjThd = NULL, valueThd = NULL, upThd = NULL, dnThd = NULL)

\S4method{movStat}{PACdataset}(object, minPAT = c(1, 5, 10), ofilePrefix = NULL)
}
\arguments{
\item{object}{Normally a movRes or PACdataset object.}

\item{...}{Arguments passing to other functions.}

\item{padjThd}{cutoff to filter rows with aMovRes@pairwise$padj<padjThd.}

\item{valueThd}{cutoff to filter rows with aMovRes@pairwise$value|>=valueThd.}

\item{upThd}{cutoff to filter up-regulated rows with aMovRes@pairwise$value>=upThd.}

\item{dnThd}{cutoff to filter down-regulated rows with aMovRes@pairwise$value<=dnThd.}

\item{minPAT}{min PAT number to filter valid PACs with expression level >= minPAT, can be like 5 or c(5,10)}

\item{ofilePrefix}{output stat results to a txt file named <ofilePrefix.patX.stat>. If it is NULL, then do not output to file.}
}
\value{
A list of several data frames including nsig, siglist, tf01, ovp, de01, and deNum.

A list of data frames representing results from different cutoffs like [PAT1, PAT5, PAT10]. Each item is a data frame recording statistical results.
}
\description{
movStat(object) makes statistics for a movRes or PACdataset object.
It is a generic function to make statistics of a object which is normally of movRes or PACddataset class.

movStat(movRes) makes statistics for a movRes object.

movStat(PACdataset) makes statistics for a PACdataset object.
}
\details{
Statistical results of movStat(movRes) are numbers of significant DE genes/PACs/switching events,
the significant gene/PAC list for each condition pair, etc.

Statistical results of movStat(PACdataset) are the distributions PATs, PACs etc.
}
\section{Methods (by class)}{
\itemize{
\item \code{movStat(movRes)}: for a movRes object.

\item \code{movStat(PACdataset)}: for a PACdataset object.

}}
\examples{
## movStat(PACdataset) ##
data(PACds)
## Stat a PACds, filtering PACs with cutoffs of 1, 5, and 10.
pstats=movStat(PACds, minPAT=c(1,5,10), ofilePrefix='statFileEach')

## If the PACds includes replicates, you may want to pool replicates first and then make stats.
pstats=movStat(subsetPACds(PACds, group='group', pool=TRUE),
              minPAT=c(1,5,10), ofilePrefix='statFilePool')
}
\seealso{
\code{\link{plotPACdsStat}} for plot the stat results of PACdataset.

Other mov-like functions: 
\code{\link{movSelect}()}
}
\concept{mov-like functions}
