% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{movUTRtrend}
\alias{movUTRtrend}
\title{Analyze 3' UTR shortening/lengthening.}
\usage{
movUTRtrend(PACds, group, method='linearTrend',
avgPACtag=5, avgGeneTag=10, aMovDEPACRes=NULL, DEPAC.padjThd=0.1,
mindist=50, fisherThd=0.1, logFCThd=1, selectOne='farest')
}
\arguments{
\item{PACds}{a PACdataset.}

\item{group}{a sample group name to get all conditions under the group.}

\item{method}{can be linearTrend, DE, DEX.}

\item{avgPACtag}{filter PACs with average tag number between cond1-cond2 >= avgPACtag.}

\item{avgGeneTag}{the same as avgPACtag but for gene.}

\item{aMovDEPACRes}{used only for method=DE/DEX, an object of movDEPACRes to get DE PAC results.}

\item{DEPAC.padjThd}{used only for method=DE/DEX, to filter DE PACs with padj<DEPAC.padjThd.}

\item{mindist, fisherThd, logFCThd}{Used to filter all switching PAC pairs.}

\item{fisherThd}{used to filter DE PAC list from DE/DEX method by pvalue<fisherThd.}

\item{logFCThd}{used to filter |rcij| >=logFCThd, where |rcij| is the log2(ratio of PA1 between two samples/ratio of PA2).}

\item{selectOne}{used only for method=DE/DEX, can be 'farest','logFC','fisherPV' to filter one switching pair.}
}
\value{
a movUTRTrendRes object, having the following slots (type, method, group, conds, pairwise=list(padj, value), fullList)
Here @fullList=[cond1.cond2, cond3.cond4, ...].
The full columns are: gene, nPAC, geneTag1, geneTag2, avgUTRlen1, avgUTRlen2, pvalue, padj,
  change <linear: cor, logRatio>, PAs1, PAs2.
}
\description{
movUTRtrend uses three methods for analysis of 3' UTR shortening/lengthening (or called 3' UTR switching).
It performs analyses for all condition pairs of the given group. If you only need a specific condition pair, call
subsetPACds first to filter used conditions.
}
\details{
For linearTrend, switching events are not per-filtered by padj value, the change column (-1 or 1) is determined by cor value.
For DE/DEX, it will first filter significant switching paris by fisherThd and logFCThd, and then select one pair by selectOne.
}
\examples{
data(PACds)
## Detect 3'UTR lengthening/shortening events using the linear trend method.
## Only PACs and genes with average read count between the two conditions >=10 and >=20 are used.
utr=movUTRtrend(PACds, group='group', method='linearTrend', avgPACtag=10, avgGeneTag=20)
## Number of genes for analyzing, including those not significant.
lapply(utr@fullList, nrow)
head(utr@fullList[["anther.embryo"]])
## Detect 3'UTR lengthening/shortening events using the DEX method.
## First get DE PAC results by DEXseq.
DEXPAC=movDEPAC(PACds, method='DEXseq', group='group', minSumPAT=10)
## Then get 3UTR switching events.
swDEX=movUTRtrend(PACds, group='group', method='DEX',
                  avgPACtag=10, avgGeneTag=20,
                  aMovDEPACRes=DEXPAC, DEPAC.padjThd=0.01,
                  mindist=50, fisherThd=0.01, logFCThd=1, selectOne='farest')
## Number of genes for analyzing, including those not significant.
lapply(swDEX@fullList, nrow)
head(swDEX@fullList[["anther.maturePollen"]])
}
\seealso{
Other comparison functions: 
\code{\link{movAPAindexDiff}()},
\code{\link{movAPAindex}()},
\code{\link{movAPAswitch}()},
\code{\link{movDEGene}()},
\code{\link{movDEPAC}()},
\code{\link{movPAindex}()},
\code{\link{plotCummPAindex}()}
}
\concept{comparison functions}
