% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\docType{class}
\name{movUTRTrendRes-class}
\alias{movUTRTrendRes-class}
\title{movUTRTrendRes object}
\description{
movUTRTrendRes is an S4 class to represent 3' UTR switching results among conditions.
}
\details{
The movUTRTrendRes object is obtained by calling movUTRtrend function.
}
\section{Slots}{

\describe{
\item{\code{fullList}}{records the full information of 3' UTR switching results,
including 3' UTR length, poly(A) sites expression (PAs1 and PAs2), etc.}
}}

\seealso{
\code{\link{movUTRtrend}} for 3' UTR switching analysis on a PACdataset.

Other movRes objects: 
\code{\link{heatmapResults-class}},
\code{\link{movAPASwitchRes-class}},
\code{\link{movDEGeneRes-class}},
\code{\link{movIndexDiffRes-class}},
\code{\link{movRes-class}}
}
\concept{movRes objects}
