% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{outputHeatStat}
\alias{outputHeatStat}
\title{Output movStat results}
\usage{
outputHeatStat(heatStats, ostatfile, plotPre, show_rownames = FALSE)
}
\arguments{
\item{heatStats}{a list from movStat().}

\item{ostatfile}{an output txt file name.}

\item{plotPre}{a file name prefix for output plots. Final output file names like <plotPre>.plots.pdf.}

\item{show_rownames}{whether show rownames in the heatmap, default is FALSE.}
}
\description{
outputHeatStat outputs statistical results from movStat() to a text file and plots figures to a pdf file.
}
\details{
This function will plot bar plots for heatStats$nsig, heatmaps for tf01, de01, deNum.
and also write statistical results to a text file.
}
\examples{
data(PACds)
swDE=movUTRtrend(PACds, group='group', method='DE',
avgPACtag=10, avgGeneTag=20,
aMovDEPACRes=DEPAC, DEPAC.padjThd=0.01,
mindist=50, fisherThd=0.01, logFCThd=1, selectOne='fisherPV')
## First call the differential analysis and then call movStat to stat the results.
stat=movStat(object=swDE, padjThd=0.01, valueThd=1)
## Output stat results into files.
## The pdf file stores the plots about the number of significant DE PACs
## and the overlap among different condition pairs.
outputHeatStat(heatStats=stat, ostatfile='3UTR_switching_DE.stat',
              plotPre='3UTR_switching_DE', show_rownames = TRUE)
}
\seealso{
Other visualization functions: 
\code{\link{movViz}()},
\code{\link{plotHeatmap}()}
}
\concept{visualization functions}
