% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_GFF.r
\name{parseGff}
\alias{parseGff}
\title{Parse gff3/gtf genome annotation}
\usage{
parseGff(aGFF)
}
\arguments{
\item{aGFF}{.gff3/.gff/.gtf file name}
}
\value{
a parsed genome annotation object, which is a list of three elements (anno.rna, anno.need, anno.frame) and can be used for annotatePAC().
}
\description{
parseGff parses genome annotation file of gff3/gtf format
}
\details{
Due to the complex GFF3/GFF/GTF structure of different genome annotation files from different species,
this function may not be always applicable for any given file. You may need to check mannually.
}
\examples{
## parse from a gff file, and save as .rda for further use.
gff=parseGff(aGFF='Bamboo.Hic.gff')
}
\seealso{
[parseTxdb()] to parse a Txdb object.

Other genome annotation functions: 
\code{\link{parseGenomeAnnotation}()}
}
\concept{genome annotation functions}
