% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_GFF.r
\name{parseTxdb}
\alias{parseTxdb}
\title{Parse TxDb genome annotation}
\usage{
parseTxdb(an.txdb)
}
\arguments{
\item{an.txdb}{a TxDb object}
}
\value{
a parsed genome annotation object, which is a list of three elements (anno.rna, anno.need, anno.frame) and can be used for annotatePAC().
}
\description{
parseTxdb parses genome annotation object of TxDb
}
\examples{
library(TxDb.Athaliana.BioMart.plantsmart28)
txdbAnno <- parseTxdb(an.txdb=TxDb.Athaliana.BioMart.plantsmart28)
}
\seealso{
[parseGff()] to parse a Gff file.
}
\concept{Genome annotation functions}
