% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{plotCummPAindex}
\alias{plotCummPAindex}
\title{Plot distribution of PA index among biological samples.}
\usage{
plotCummPAindex(
  PAindex,
  groupName = "Sample",
  xlab,
  ylab = "Cumulative fraction"
)
}
\arguments{
\item{PAindex}{a matrix of PAindex, which is normally from movAPAindex() or movPAindex().}

\item{groupName}{shown as the legend title.}

\item{xlab}{x label.}

\item{ylab}{y label, default is "Cumulative fraction".}
}
\value{
A ggplot2 plot object.
}
\description{
PlotCummPAindex plots distribution of PA index among biological samples by calculating the cumulative fraction.
}
\examples{
data(scPACds)
scPACdsCt=subsetPACds(scPACds, group='celltype', pool=TRUE)
scPACdsCt=get3UTRAPAds(scPACdsCt, sortPA=TRUE, choose2PA='PD')
## Cummlative plot of GPI index.
gpi=movAPAindex(scPACdsCt, method="GPI")
plotCummPAindex(PAindex=gpi, groupName='cell type', xlab='GPI')
## The plot of the distribution of Shannon index (tissue-specificity) for each PAC.
shan=movPAindex(scPACdsCt, method="shan")
plotCummPAindex(PAindex=shan[, -c(1:3)], groupName='cell type', xlab='Shannon index')
}
\seealso{
Other comparison functions: 
\code{\link{movAPAindexDiff}()},
\code{\link{movAPAindex}()},
\code{\link{movAPAswitch}()},
\code{\link{movDEGene}()},
\code{\link{movDEPAC}()},
\code{\link{movPAindex}()},
\code{\link{movUTRtrend}()}
}
\concept{comparison functions}
