% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{readDESeqDs}
\alias{readDESeqDs}
\title{Read a DESeqDataset}
\usage{
readDESeqDs(datafile, annofile, filetype, noIntergenic = TRUE, PAname = NULL)
}
\arguments{
\item{datafile}{a gene/PA file.}

\item{annofile}{annotation for samples.}

\item{filetype}{specify the file type, gene/PA.}

\item{noIntergenic}{valid when filetype=PA, specifying whether to remove intergenic PACs.}

\item{PAname}{valid when filetype=PA, specifying how to set the name (rowname) of PACs.
PAname=PA, the PA name is set as 'gene.PAN';
PAname=coord, then 'gene.coord'.}
}
\value{
A DESeqDataset with each row a gene or PAC.
}
\description{
readDESeqDs reads file of PACs or genes into a DESeqDataset.
}
\examples{
data(PACds)
writePACds(PACds, file='pacds.txt', colDataFile='coldata.txt')

## Read PACs to DESeqDataset
dds=readDESeqDs(datafile='pacds.txt', annofile='coldata.txt',
     filetype='PA',noIntergenic=TRUE, PAname='PA')
head(mcols(dds)); colData(dds); head(assays(dds)$counts)
}
