% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{removePACdsIP}
\alias{removePACdsIP}
\title{Remove internal priming.}
\usage{
removePACdsIP(PACds, bsgenome, returnBoth=TRUE, up=-10, dn=10, conA=6, sepA=7)
}
\arguments{
\item{PACds}{a PACdataset.}

\item{bsgenome}{a BSGenome/FaFile to get chromosome sequences.
The PACds@anno$chr should all be in bsgenome$seqnames.}

\item{returnBoth}{If TRUE, then return a list(real, ip), otherwise return only the real PACds.}

\item{up}{the start position of the window to scan, see faFromPACds().}

\item{dn}{the end position of the window to scan, see faFromPACds().
If up=-10 and dn=10 then will scan the [-10, 0, +10] window (21 nt).
If up=-10 and dn=-1 then will scan the [-10, -1] window (10 nt, not including the PA pos).}

\item{conA}{the number of continuous As in a window to define IP. Default is NA.}

\item{sepA}{the number of total As in a window to define IP. Default is NA, which means not considering total As.}

\item{ipGrams}{the grams to define IP. Default is NA. If any one gram in ipGrams is located within [up, dn] region then it is IP.}

\item{chrCheck}{if TRUE, then all chr in PACds should be in bsgenome, otherwise will ignore those non-consistent chr rows in PACds.}
}
\value{
Dependent on returnBoth.
}
\description{
removePACdsIP removes internal priming artificats from a PACdataset object.
}
\details{
If there are continuous [conA] As or at least >=[sepA] As or at least one [ipGrams] in the [up, dn] window of a PAC, then it is considered as IP.
}
\examples{
data(PACds)
library("BSgenome.Oryza.ENSEMBL.IRGSP1")
bsgenome <- BSgenome.Oryza.ENSEMBL.IRGSP1
## remove internal priming within -140 ~ +10 of PA, without considering total A#
removePACdsIP(PACds, bsgenome, returnBoth=TRUE, up=-140, dn=10, conA=6, sepA=NA)
## remove internal priming within -10 ~ +10 of PA, if there is any AAAAAA or total 7 As.
removePACdsIP(PACds, bsgenome, returnBoth=TRUE, up=-10, dn=10, conA=6, sepA=7)
## remove internal priming within -10 ~ +10 of PA, if there is any one ipGrams.
removePACdsIP(PACds, bsgenome, returnBoth=TRUE, up=-5, dn=10, ipGrams=c('AAAA', 'AGAA', 'AAGA', 'AAAG'))
}
\seealso{
Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByKnownPAC}()},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{countOvpPACds}()},
\code{\link{ext3UTRPACds}()},
\code{\link{findOvpPACds}()},
\code{\link{get3UTRAPAds}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{samplePACds}()},
\code{\link{scPACds}},
\code{\link{subsetPACds}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
