% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{samplePACds}
\alias{samplePACds}
\title{Sample a PACdataset}
\usage{
samplePACds(pacds, N, nPAT = 1)
}
\arguments{
\item{pacds}{a PACdataset}

\item{N}{number of sampled PACs. If N>nrow(pacds), then do not sampling.}

\item{nPAT}{filter PACs with total tagnum>=nPAT before sampling.}
}
\value{
A subset of PACdataset.
}
\description{
samplePACds returns a random subset of PACdataset.
}
\examples{
data(PACds)
pacds=samplePACds(PACds, N=1000)
}
\concept{PACdataset functions
\code{\link{subsetPACds}} to get a subset PACdataset.}
