% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\docType{data}
\name{scPACds}
\alias{scPACds}
\title{A demo PACdataset of mouse sperm cells}
\format{
A PACdataset with 771 PACs from single cells.
}
\source{
<Shulman et al, 2019. Nucleic Acids Res>
}
\usage{
scPACds
}
\description{
A dataset of the PACdataset class from mouse sperm cells, containing 771 PACs from 396 genes located in chromosome 12.
There are total 2042 cells from three cell types (SC, Spermatocytes; RS, Round spermatids; ES, Elongating spermatids).
This dataset contains the gene Psen1 (ENSMUSG00000019969) presented in Shulman et al, 2019.
The original PAC dataset was obtained from Shulman et al, 2019, but re-annotated by the MM10 genome annotation (TxDb.Mmusculus.UCSC.mm10.ensGene).
}
\seealso{
Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByKnownPAC}()},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{countOvpPACds}()},
\code{\link{ext3UTRPACds}()},
\code{\link{findOvpPACds}()},
\code{\link{get3UTRAPAds}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{samplePACds}()},
\code{\link{subsetPACds}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
\keyword{datasets}
