% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{subscript_operator}
\alias{subscript_operator}
\title{Subset rows and/or columns of a PACdataset}
\usage{
\S4method{[}{PACdataset,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{a PACdataset}
}
\value{
A new PACdataset.
}
\description{
Subset rows of a PACdataset, by overloading of subscript operator.
}
\examples{
data(PACds)
PACds[1:100, 2:5]
PACds[1:100, ]
PACds[, 1:5]
PACds[, c(T,T,F)]
PACds[, c('anther1','anther2')]
}
\seealso{
Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{createPACdataset}()},
\code{\link{get3UTRAPAds}()},
\code{\link{get3UTRAPApd}()},
\code{\link{length}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{scPACds}},
\code{\link{summary}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
