% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{subsetPACds}
\alias{subsetPACds}
\title{Subset a PACdataset}
\usage{
subsetPACds(
  pacds,
  group = NULL,
  cond1 = NULL,
  cond2 = NULL,
  conds = NULL,
  avgPACtag = 0,
  avgGeneTag = 0,
  totPACtag = 0,
  choosePA = NULL,
  PAs = NULL,
  genes = NULL,
  chrs = NULL,
  noIntergenic = FALSE,
  avg = FALSE,
  pool = FALSE,
  clearPAT = 0,
  verbose = FALSE
)
}
\arguments{
\item{group}{a sample group, must be present in PACds@colData.
If NULL, then parameters of cond1, cond2, and avg are not in use.}

\item{cond1}{a condition, must be present in PACds@colData.}

\item{cond2}{a condition, must be present in PACds@colData.
If cond1 or cond2 is NULL, then both cond1 and cond2 are set NULL.}

\item{conds}{to subset multiple conditions.}

\item{avgPACtag}{if >0, then filter PACs with average PAT number across all samples >= avgPACtag, after subseting by group and conds.}

\item{avgGeneTag}{similar to avgPACtag, but for gene.}

\item{totPACtag}{filter PACs with total PAT number of all samples >=totPACtag.}

\item{choosePA}{value can e NULL/distal/proximal/APA. Only if PACds is all 3UTRAPA, then choosePA can be distal/proximal.
APA means choose PACs from APA genes.}

\item{PAs}{to filter PACs by a PAC name list according to rownames in PAs.}

\item{genes}{to filter by a gene name list according to PACds@anno$gene.}

\item{chrs}{to filter by a chr name list according to PACds@anno$chr.}

\item{noIntergenic}{whether to remove intergenic PACs.}

\item{avg}{TRUR/FALSE. If TRUE then take average of replicates of each of subset conditions.}

\item{pool}{TRUR/FALSE. If TRUE then take pool value of replicates of each of subset conditions.}

\item{clearPAT}{If >0 then set @counts[<clearPAT]=0 and remove blank lines and reset PACds@anno.}
}
\value{
a subset of PACdataset with 0-count lines removed.
}
\description{
subsetPACds returns a subset of PACdataset.
}
\examples{
data(PACds)
PACds1=subsetPACds(PACds, group='group', pool=TRUE)
## Subset two conditions.
pacds=subsetPACds(PACds, group='group',cond1='anther', cond2='embryo')
## Subset PACs from APA genes in two conditions.
PACds1=subsetPACds(PACds, group='group', cond1='anther', cond2='embryo', choosePA='apa')
## Subset PACs of given genes.
subsetPACds(PACds, genes=PACds@anno$gene[1:5], verbose=TRUE)
## Subset PACs of given PA names.
swPAC=subsetPACds(PACds, PAs=rownames(PACds@counts)[1:5], verbose=TRUE)
}
\seealso{
[samplePACds()] to get sampled PACds.

Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByKnownPAC}()},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{countOvpPACds}()},
\code{\link{ext3UTRPACds}()},
\code{\link{findOvpPACds}()},
\code{\link{get3UTRAPAds}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{samplePACds}()},
\code{\link{scPACds}},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
