% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{testExt3UTR}
\alias{testExt3UTR}
\title{Test extended 3'UTR length}
\usage{
testExt3UTR(pacds, ext3UTRlen = seq(1000, 10000, by = 1000), toPlot = TRUE)
}
\arguments{
\item{pacds}{a PACdataset object, which should already be annotated by annotatePAC().}

\item{ext3UTRlen}{an integer vector denoting the extended lengthes, like c(1000, 2000, 3000)}

\item{toPlot}{whether to plot a barplot for the number changes.}
}
\value{
A data frame with two columns: ext3UTRLen, addedPAnum denoting numbers of recruited 3UTR PAs for each extended length.
}
\description{
testExt3UTR test the number of 3UTR PA changes after extending a given length.
}
\examples{
data(PACds)
testExt3UTR(PACds, seq(1000, 10000, by=1000))
}
\concept{PACdataset functions
\code{\link{ext3UTRPACds}} to extend3UTR for a PACdataset.}
