% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{writePACds}
\alias{writePACds}
\title{Write a PACdataset into a file}
\usage{
writePACds(PACds, file, colDataFile = NULL)
}
\arguments{
\item{PACds}{a PACdataset object.}

\item{file}{a file name.}

\item{colDataFile}{a file name. If not NULL, then output PACds@colData into this file.}
}
\value{
NULL.
}
\description{
writePACds writes a PACdataset object into a file.
}
\examples{
data(PACds)
writePACds(PACds, file='pacds.txt', colDataFile='coldata.txt')
readPACds(pacFile='pacds.txt', colDataFile='coldata.txt')
}
\seealso{
Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{createPACdataset}()},
\code{\link{get3UTRAPAds}()},
\code{\link{get3UTRAPApd}()},
\code{\link{length}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{scPACds}},
\code{\link{subscript_operator}},
\code{\link{summary}()}
}
\concept{PACdataset functions}
