% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{kcount}
\alias{kcount}
\title{Count kgrams}
\usage{
kcount(seq=NULL, fafile=NULL, grams=NULL, k=NULL, kfile=NULL, from=NA, to=NA, sort=TRUE, topn=50)
}
\arguments{
\item{seq}{a DNAStringSet object.}

\item{fafile}{fasta file name(s). fafile or seq can only set one.
If fafile=c(A='fA1.fa',B='fB2.fa'), then the output depends on mergeFile=TRUE/FALSE. And the column names are set as A for count, A_perc for percent.
If names(fafile) is not provided, then output columns will be A1_perc, B2_perc.}

\item{grams}{specify the grams to count. Value can be c('aataa',...) or 'V1' (means AATAAA and its 1nt variants) or mouse/mm/mm10 (means PAS of mouse).}

\item{k}{specify k grams. E.g., k=6 means all hexamers.}

\item{kfile}{specify a kgram file, with each line being a kgram. The file does not contain header line.
Only one parameter of grams/k/kfile can be provided.}

\item{mergeFile}{if TRUE, then when multiple file names are given by fafile, files are combined first.}

\item{sort}{TRUE/FALSE, whether to sort kgrams by their counts.}

\item{topn}{whether to output topn grams.}

\item{perc}{whether also to output percentage.}

\item{from/to}{subset seq first before counting. If a sequence is shorter, then the sequence will be deleted.}
}
\value{
a data frame with columns [grams, count, <perc>] or [grams, file1,..., fileN, ..., file1_perc, ..., fileN_perc]
}
\description{
kcount counts given kgrams from given sequences. Kgrams can be specified by character vector or by K=N.
}
\examples{
kcount(seq=seq, grams=c('AATAAA','ATTAAA'))
kcount(seq=seq, k=3)
kcount(seq=seq, k=6, from=260, to=290)
kcount(fafile='updn.fa', k=6, from=260, to=290)
kcount(fafile='updn.fa', grams='v1', from=260, to=290)
kcount(fafile='updn.fa', grams='mouse', from=240, to=310)
kcount(fafile=c('updn.fa', 'updn2.fa'), grams='mouse', from=240, to=310, mergeFile=FALSE)
kcount(fafile=c(A='updn.fa', B='updn2.fa'), grams='mouse', from=240, to=310, mergeFile=FALSE)
kcount(fafile=c('updn.fa', 'updn2.fa'), grams='mouse', from=240, to=310, mergeFile=TRUE)
}
\seealso{
Other APA signal functions: 
\code{\link{annotateByPAS}()},
\code{\link{faFromPACds}()},
\code{\link{getVarGrams}()},
\code{\link{plotATCGforFAfile}()},
\code{\link{plotSeqLogo}()}
}
\concept{APA signal functions}
