% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{mergePACds}
\alias{mergePACds}
\title{Merge multiple PACdatasets}
\usage{
mergePACds(PACdsList, d=24)
}
\arguments{
\item{PACdsList}{a PACdataset, or a list of multiple PACdataset objects. The PACds@anno should have columns chr/strand/coord.
If there is no colData in PACds, then the sample label will be set as groupN.
If PACdsList is a PACdataset, then will treat it as PA and group nearby PAs into PACs.}

\item{d}{distance to group nearby PACds, default is 24 nt.}
}
\value{
A merged PACdataset. The counts slot stores counts of merged samples.
If sample names from different PACdataset objects are duplicated, then will be set as .x,.y.
The colData slot stores the merge sample annotation from the first column of each @colData.
The anno slot contains these columns: chr, strand, coord, tottag, UPA_start, UPA_end, nPA, maxtag.
}
\description{
mergePACds groups nearby PACs from single/multiple PACdataset objects.
}
\details{
This function is particularlly useful for grouping nearby cleavage sites into PACs.
It is also useful When you have multiple PA or PAC files, each file is from one sample.
Then you need to merge these PACds into one PACds for DE or other analyses.
But after grouping and/or merging, you may need call annotatePAC to annotate the merged PACs by a GFF annotation.
}
\examples{
## Group PA into PACs
data(PACds)
PACds@counts=rbind(PACds@counts, PACds@counts)
PACds@anno=rbind(PACds@anno, PACds@anno)
ds=mergePACds(PACds, d=24)
## merge two PACds
ds1=makeExamplePACds()
ds2=makeExamplePACds()
ds=mergePACds(list(ds1, ds2), d=24)
}
\seealso{
[annotatePAC()] to annotate a PACdataset; [rbind()] to combine multiple PACdatasets of the same format.

Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByKnownPAC}()},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{countOvpPACds}()},
\code{\link{ext3UTRPACds}()},
\code{\link{findOvpPACds}()},
\code{\link{get3UTRAPAds}()},
\code{\link{makeExamplePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{samplePACds}()},
\code{\link{scPACds}},
\code{\link{subsetPACds}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
