% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{movAPAindex}
\alias{movAPAindex}
\title{Get APA index}
\usage{
movAPAindex(PACds, method, choose2PA=NULL, RUD.includeNon3UTR=FALSE, clearPAT=0)
}
\arguments{
\item{PACds}{a PACdataset.}

\item{method}{Can be WUL, RUD, SLR.
WUL: Weighted UTR length (Jia 2017; Ultisky) (3UTRAPA, choose2PA=NULL/PD/MOST)
RUD: Relative Usage of Distal Poly(A) Sites (Ji 2009) (3UTRAPA with/without non-3UTR PACs, choose2PA=NULL/PD/MOST) only calc distal ratio, RUD=distal/gene
SLR: short to long ratio (Begik 2017) (3UTRAPA, choose2PA=PD/most) #only for PD, SLR=short/long
GPI: geometric proximal index (Shulman 2019) (3UTRAPA, choose2PA=NULL/PD/most) #GPI=proximal/geo_mean_PD}

\item{choose2PA}{specify whether and how to choose two PACs. The value can be NULL, PD (choose only proximal and distal sites),
farest (choose two PACs that are with the longest distance), most (choose two PACs with the most abundance).}

\item{RUD.includeNon3UTR}{only used when method=RUD, to use all PA or just 3UTR PA.}

\item{clearPAT}{set PAC counts to 0 when it is < clearPAT.
This option is used to remove the effect of low expressed PACs. Otherwise, PAC with tag=1 may have high impact on RUD/SLR calculations.
After clearPAT, some results will be set as NaN.}

\item{group}{a sample group name to get all conditions under the group.}
}
\value{
A data frame with rownames denoting genes, columns are samples, and values are the APA index values. Also includes other method-specific columns.
}
\description{
movAPAindex calculates APA index by different metrics for each condition.
}
\details{
This funtion uses all samples to filter APA sites, and then gets index for each sample.
So there might be NaN for some samples without enough PAC (tag=0).
The final result may have NaN but will not have Inf (it has been assigned the max value for Inf element).
For method=WUL, NaN happens when the PAC expression in sample A for gene X is 0.
For method=RUD (distal/gene), NaN happens when the PAC expression in sample A for gene X is 0.
For method=SLR (short/long), NaN happens when both the proximal and distal PAC is 0; Inf happens when distal PAC=0 but proximal not 0.
}
\examples{
data(PACds)
gs=movAPAindex(PACds, method="WUL", choose2PA=NULL)
gs=movAPAindex(PACds, method="SLR", choose2PA='PD', clearPAT=5)

## GPI index for only proximal PAC
geneGPI2=movAPAindex(PACds, method="GPI", choose2PA=NULL) #use all 3UTR PACs

geneGPI=movAPAindex(PACds, method="GPI", choose2PA='PD') #use only proximal&distal PACs
PACds1=get3UTRAPAds(PACds, sortPA=TRUE, choose2PA='PD')
geneGPI1=movAPAindex(PACds1, method="GPI", choose2PA='PD')
identical(geneGPI1, geneGPI) ## SAME
## use movPAindex to get GPI (but also including distal GPI)
paGeo=movPAindex(PACds1, method='geo')
}
\seealso{
Other comparison functions: 
\code{\link{movAPAindexDiff}()},
\code{\link{movAPAswitch}()},
\code{\link{movDEGene}()},
\code{\link{movDEPAC}()},
\code{\link{movPAindex}()},
\code{\link{movUTRtrend}()},
\code{\link{plotCummPAindex}()}
}
\concept{comparison functions}
