% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\docType{class}
\name{movDEPACRes-class}
\alias{movDEPACRes-class}
\title{movDEPACRes object}
\description{
movUTRTrendRes is an S4 class to represent DE PAC (poly(A) site cluster) results among conditions.
}
\details{
The movDEPACRes object is obtained by calling movDEPAC function.
}
\section{Slots}{

\describe{
\item{\code{PACusage}}{When method=DEXseq, PACusage slot stores the exon (PAC) usage from DEXseq.}
}}

\seealso{
[movDEPAC()] for DE PAC analysis on a PACdataset.

Other movRes objects: 
\code{\link{heatmapResults-class}},
\code{\link{movAPASwitchRes-class}},
\code{\link{movDEGeneRes-class}},
\code{\link{movIndexDiffRes-class}},
\code{\link{movRes-class}},
\code{\link{movUTRTrendRes-class}}
}
\concept{movRes objects}
