% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{normalizePACds}
\alias{normalizePACds}
\title{Normalize a PACdataset}
\usage{
normalizePACds(pacds, method='TPM')
}
\arguments{
\item{pacds}{a PACdataset.}

\item{method}{normalization method, can be TPM/CPM/DEseq/DESeq2/EdgeR.
CPM=TPM (after normalization, libsize of all columns are 1e6), DESeq=DESeq2.}
}
\value{
a normalized PACdataset with @counts normalized.
}
\description{
normalizePACds returns a normalized PACds with @counts normalized.
}
\details{
This function is used for normalization with different methods.
}
\examples{
data(PACds)
PACdsNorm=normalizePACds(PACds, method='TPM')
}
\seealso{
Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByKnownPAC}()},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{countOvpPACds}()},
\code{\link{ext3UTRPACds}()},
\code{\link{findOvpPACds}()},
\code{\link{get3UTRAPAds}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{samplePACds}()},
\code{\link{scPACds}},
\code{\link{subsetPACds}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
