% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\docType{data}
\name{PACds}
\alias{PACds}
\title{A demo PACdataset of rice tissues}
\format{
A PACdataset with 1233 PACs.
}
\source{
<Zhu et al, 2019. Plant Physiol>
}
\usage{
PACds
}
\description{
A dataset of the PACdataset class from rice japonica, containing 1233 PACs in 455 genes from three tissues
Biological samples are anther, embryo, and maturePollen, each with three replicates.
PAC counts have been normalized.
}
\seealso{
Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{annotateByKnownPAC}()},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{countOvpPACds}()},
\code{\link{ext3UTRPACds}()},
\code{\link{findOvpPACds}()},
\code{\link{get3UTRAPAds}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{rbind}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{samplePACds}()},
\code{\link{scPACds}},
\code{\link{subsetPACds}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
\keyword{datasets}
