% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_funclib_movAPA.r
\name{rbind}
\alias{rbind}
\title{Combine multiple PACdatasets by rows}
\usage{
rbind(...)
}
\arguments{
\item{...}{one or more PACdataset object.}
}
\value{
If params are PACdatasets with the same column definations, then will combine rows of PACds. See ?base::cbind for the value returned by the default methods.
}
\description{
Combine multiple PACdatasets by rows
}
\seealso{
Other PACdataset functions: 
\code{\link{PACdataset-class}},
\code{\link{PACds}},
\code{\link{annotateByPAS}()},
\code{\link{annotatePAC}()},
\code{\link{createPACdataset}()},
\code{\link{get3UTRAPAds}()},
\code{\link{get3UTRAPApd}()},
\code{\link{length}()},
\code{\link{makeExamplePACds}()},
\code{\link{mergePACds}()},
\code{\link{normalizePACds}()},
\code{\link{plotPACdsStat}()},
\code{\link{readPACds}()},
\code{\link{removePACdsIP}()},
\code{\link{scPACds}},
\code{\link{subscript_operator}},
\code{\link{summary}()},
\code{\link{writePACds}()}
}
\concept{PACdataset functions}
